requirements.txt
torch
torchvision
numpy
opencv-python
matplotlib
pandas
scikit-image
tqdm
pillow
